+++ src/service/layers/infrastructure/repositories/mongoose/types.ts
@@ -1,53 +1,17 @@
-import type {
-  Document,
-  MongooseModel,
-  Schema,
-  Default__v,
-  Require_id,
-  ObjectId,
-  IfAny,
-} from '../../../../../third-party/mongoose';
+import type { MongooseModel, Schema } from '../../../../../third-party/mongoose';
+import type { ORMType } from '../../orm/types';
 
-export enum ORMType {
-  Mongoose = 'mongoose',
-  Sequelize = 'sequelize',
-}
-
-export type WithMongooseMeta<T> = Default__v<Require_id<T>>;
-
-export type MongooseDocument<Db> = Document<
-  ObjectId,
-  {}, // eslint-disable-line @typescript-eslint/no-empty-object-type
-  Db,
-  Record<string, string>
-> &
-  WithMongooseMeta<Db>;
-
-export type LeanWithMeta<T> = WithMongooseMeta<T> & {
-  id: string;
-  createdAt: Date;
-  updatedAt: Date;
-};
-
-export type MongooseDoc<Db> = IfAny<
-  MongooseDocument<Db>,
-  MongooseDocument<Db>,
-  MongooseDocument<Db>
->;
-
-export type ToLeanInput<D, T> =
-  | MongooseDoc<D>
-  | (T & {
-      id: string;
-      createdAt: Date;
-      updatedAt: Date;
-    })
-  | null
-  | undefined;
+export type {
+  WithMongooseMeta,
+  LeanDocument,
+  LeanWithMeta,
+  MongooseDoc,
+  ToLeanInput,
+} from '../../orm/mongoose/types';
 
 export interface MongooseRepositoryConfig<D = unknown> {
   readonly type?: ORMType.Mongoose;
-  readonly name?: string;
+  readonly name: string;
   readonly schema: Schema<D>;
 }